refactor: cleanup, bin to lib
