+++ README.md
@@ -1,8 +1,8 @@
 # Silt-Lua, a lua subset interpreter in 100% rust
 
-This project exists to answer a problem with the current rust landscape in lacking a complete lua interpreter solution written in 100% safe rust. The existing implementations as of 2023 are missing crucial features and optmiziations to satisfy the requirements of my closed source project [Petrichor64](https://makeavoy.itch.io/petrichor64), so here we are.
+This project exists to answer a problem with the current rust landscape in lacking a complete lua interpreter solution written in 100% rust. The existing implementations as of 2023 are missing crucial features and optmiziations to satisfy the requirements of my closed source project [Petrichor64](https://makeavoy.itch.io/petrichor64), so here we are.
 
-Core focus is a basic interpreter with minor overhead and reletive speed comparison to standard lua. The goal is for perfect wasm target building. Userdata will mirror traits used by the mlua and rlua crates to allow easy drop in.
+Core focus is a basic interpreter with minor overhead and reletive speed comparison to standard lua. The goal is for a perfect wasm32-unknown-unknown target build. No emscripten necessary! Userdata will mirror traits used by the mlua and rlua crates to allow easy drop in.
 
 Secondary goals are CLI, LSP, and full standard library compliance. Ideally loading third party libraries as well, if under the silt subset then the ability to import libraries as pure lua will be added, in case of any potential overlooked token conflicts. Again, these goals are TBD and not top priority.
 
@@ -16,3 +16,9 @@ This has been written from the ground up with observations of the lua langauge a
 - anoynmous arrow functions of the -> (C# style) are supported `func_name =param -> param+1` in addition to this functons will return the last statement value
 - bang usage ! for not or not equal (~=) can be used if you're hard pressed to not use them like I am. They do not replace not or ~=, only act as builtin aliases
 - numbers can include underscores which are ignored characters used for readability, borrowed right from rust
+
+## Examples
+
+```rust
+let test= 5*5;
+```
